+++ R/Connect.R
@@ -49,8 +49,7 @@
 #' connectionDetails <- createConnectionDetails(dbms = "postgresql",
 #'                                              server = "localhost/postgres",
 #'                                              user = "root",
-#'                                              password = "blah",
-#'                                              schema = "cdm_v4")
+#'                                              password = "blah")
 #' conn <- connect(connectionDetails)
 #' dbGetQuery(conn, "SELECT COUNT(*) FROM person")
 #' disconnect(conn)
@@ -66,16 +65,9 @@ createConnectionDetails <- function(dbms,
                                     oracleDriver = "thin",
                                     connectionString = NULL,
                                     pathToDriver = getOption("pathToDriver")) {
-  # First: get default values:
   result <- list()
   for (name in names(formals(createConnectionDetails))) {
     result[[name]] <- get(name)
-  }
-  # Second: overwrite defaults with actual values:
-  values <- lapply(as.list(match.call())[-1], function(x) eval(x, envir = sys.frame(-3)))
-  for (name in names(values)) {
-    if (name %in% names(result))
-      result[[name]] <- values[[name]]
   }
   class(result) <- "connectionDetails"
   return(result)
@@ -163,12 +155,11 @@ findPathToJar <- function(name, pathToDriver) {
 #' conn <- connect(dbms = "postgresql",
 #'                 server = "localhost/postgres",
 #'                 user = "root",
-#'                 password = "xxx",
-#'                 schema = "cdm_v4")
+#'                 password = "xxx")
 #' dbGetQuery(conn, "SELECT COUNT(*) FROM person")
 #' disconnect(conn)
 #'
-#' conn <- connect(dbms = "sql server", server = "RNDUSRDHIT06.jnj.com", schema = "Vocabulary")
+#' conn <- connect(dbms = "sql server", server = "RNDUSRDHIT06.jnj.com")
 #' dbGetQuery(conn, "SELECT COUNT(*) FROM concept")
 #' disconnect(conn)
 #'
@@ -176,7 +167,6 @@ findPathToJar <- function(name, pathToDriver) {
 #'                 server = "127.0.0.1/xe",
 #'                 user = "system",
 #'                 password = "xxx",
-#'                 schema = "test",
 #'                 pathToDriver = "c:/temp")
 #' dbGetQuery(conn, "SELECT COUNT(*) FROM test_table")
 #' disconnect(conn)
@@ -199,6 +189,12 @@ connect <- function(connectionDetails = NULL,
                     oracleDriver = "thin",
                     connectionString = NULL,
                     pathToDriver = getOption("pathToDriver")) {
+  
+  if (!missing(schema) && !is.null(schema)) {
+    warning(
+      "The schema argument is deprecated. Please use fully specified table names in your SQL statement, for example 'SELECT * FROM my_schema.my_table;'"
+    )
+  }
   if (!missing(connectionDetails) && !is.null(connectionDetails)) {
     connection <- connect(dbms = connectionDetails$dbms,
                           user = connectionDetails$user,
@@ -214,7 +210,7 @@ connect <- function(connectionDetails = NULL,
     return(connection)
   }
   if (dbms == "sql server") {
-    jarPath <- findPathToJar("^sqljdbc.*\\.jar$", pathToDriver)
+    jarPath <- findPathToJar("^mssql-jdbc.*.jar$|^sqljdbc.*\\.jar$", pathToDriver)
     driver <- getJbcDriverSingleton("com.microsoft.sqlserver.jdbc.SQLServerDriver", jarPath)
     if (missing(user) || is.null(user)) {
       # Using Windows integrated security
@@ -254,7 +250,7 @@ connect <- function(connectionDetails = NULL,
   }
   if (dbms == "pdw") {
     writeLines("Connecting using SQL Server driver")
-    jarPath <- findPathToJar("^sqljdbc.*\\.jar$", pathToDriver)
+    jarPath <- findPathToJar("^mssql-jdbc.*.jar$|^sqljdbc.*\\.jar$", pathToDriver)
     driver <- getJbcDriverSingleton("com.microsoft.sqlserver.jdbc.SQLServerDriver", jarPath)
     if (missing(user) || is.null(user)) {
       # Using Windows integrated security
@@ -398,7 +394,11 @@ connect <- function(connectionDetails = NULL,
   if (dbms == "redshift") {
     writeLines("Connecting using Redshift driver")
     jarPath <- findPathToJar("^RedshiftJDBC.*\\.jar$", pathToDriver)
-    driver <- getJbcDriverSingleton("com.amazon.redshift.jdbc4.Driver", jarPath)
+    if (grepl("RedshiftJDBC42", jarPath)) {
+      driver <- getJbcDriverSingleton("com.amazon.redshift.jdbc42.Driver", jarPath)
+    } else {
+      driver <- getJbcDriverSingleton("com.amazon.redshift.jdbc4.Driver", jarPath)
+    }
     if (missing(connectionString) || is.null(connectionString)) {
       if (!grepl("/", server))
         stop("Error: database name not included in server string but is required for Redshift Please specify server as <host>/<database>")
@@ -491,6 +491,26 @@ connect <- function(connectionDetails = NULL,
     attr(connection, "dbms") <- dbms
     return(connection)
   }
+  if (dbms == "hive") {
+      writeLines("Connecting using Hive driver")
+      jarPath <- findPathToJar("^hive-jdbc-standalone\\.jar$", pathToDriver)
+      driver <- getJbcDriverSingleton("org.apache.hive.jdbc.HiveDriver", jarPath)
+  
+      if (missing(connectionString) || is.null(connectionString)) {
+          connectionString <- paste0("jdbc:hive2://", server, ":", port, "/", schema)
+          if (!missing(extraSettings) && !is.null(extraSettings)) {
+              connectionString <- paste0(connectionString, ";", extraSettings)
+          }
+      }
+      connection <- connectUsingJdbcDriver(driver,
+      connectionString,
+      user = user,
+      password = password,
+      dbms = dbms)
+  
+      attr(connection, "dbms") <- dbms
+      return(connection)
+  }
   if (dbms == "bigquery") {
     writeLines("Connecting using BigQuery driver")
     
@@ -557,6 +577,10 @@ connectUsingJdbcDriver <- function(jdbcDriver,
                     stringQuote = stringQuote,
                     dbms = dbms,
                     uuid = uuid)
+  if (dbms == "hive") {
+    attr(connection, "url") <- url
+    attr(connection, "user") <- properties$user
+  }
   registerWithRStudio(connection)
   return(connection)
 }