+++ R/InsertTable.R
@@ -52,7 +52,7 @@ mergeTempTables <- function(connection, tableName, varNames, sourceNames, distri
                sep = "")
   sql <- SqlRender::translate(sql, targetDialect = connection@dbms, oracleTempSchema = oracleTempSchema)
   executeSql(connection, sql, progressBar = FALSE, reportOverallTime = FALSE)
-  
+
   # Drop source tables:
   for (sourceName in sourceNames) {
     sql <- paste("DROP TABLE", sourceName)
@@ -91,15 +91,19 @@ castValues <- function(values, fts) {
         sep = "")
 }
 
-formatRow <- function(data, castValues, fts) {
+formatRow <- function(data, aliases = c(), castValues, fts) {
   if (castValues) {
     data <- castValues(data, fts)
   }
-  return(paste(data, collapse = ","))
+  return(paste(data, aliases, collapse = ","))
 }
 
 ctasHack <- function(connection, qname, tempTable, varNames, fts, data, progressBar, oracleTempSchema) {
-  batchSize <- 1000
+  if (attr(connection, "dbms") == "hive") {
+    batchSize <- 750
+  } else {
+    batchSize <- 1000
+  }
   mergeSize <- 300
   
   if (any(tolower(names(data)) == "subject_id")) {
@@ -125,16 +129,18 @@ ctasHack <- function(connection, qname, tempTable, varNames, fts, data, progress
       tempNames <- c(mergedName)
     }
     end <- min(start + batchSize - 1, nrow(data))
-    batch <- toStrings(data[start:end, , drop = FALSE], fts)    
-
+    batch <- toStrings(data[start:end, , drop = FALSE], fts)
+    
+    varAliases <- strsplit(varNames, ",")[[1]]
     # First line gets type information:
-    valueString <- formatRow(batch[1, , drop = FALSE], castValues = TRUE, fts = fts)
+    valueString <- formatRow(batch[1, , drop = FALSE], varAliases, castValues = TRUE, fts = fts)
     if (end > start) {
       # Other lines only get type information if BigQuery:
       valueString <- paste(c(valueString, apply(batch[2:nrow(batch), , drop = FALSE],
                                                 MARGIN = 1,
                                                 FUN = formatRow,
-                                                castValues = attr(connection, "dbms") == "bigquery",
+                                                aliases = varAliases,
+                                                castValues = attr(connection, "dbms") %in% c("bigquery", "hive"),
                                                 fts = fts)), 
                            collapse = "\nUNION ALL\nSELECT ")
     }
@@ -163,26 +169,26 @@ ctasHack <- function(connection, qname, tempTable, varNames, fts, data, progress
 
 is.bigint <- function(x) {
   num <- 2^63
-  
+
   bigint.min <- -num
   bigint.max <- num - 1
-  
+
   return(!all(is.na(x)) && is.numeric(x) && !is.factor(x) && all(x == round(x), na.rm = TRUE) &&  all(x >= bigint.min, na.rm = TRUE) && all(x <= bigint.max, na.rm = TRUE))
 }
 
 #' Insert a table on the server
 #'
 #' @description
-#' This function sends the data in a data frame or ffdf to a table on the server. Either a new table
+#' This function sends the data in a data frame to a table on the server. Either a new table
 #' is created, or the data is appended to an existing table.
 #'
 #' @param connection          The connection to the database server.
 #' @param tableName           The name of the table where the data should be inserted.
-#' @param data                The data frame or ffdf containing the data to be inserted.
+#' @param data                The data frame containing the data to be inserted.
 #' @param dropTableIfExists   Drop the table if the table already exists before writing?
 #' @param createTable         Create a new table? If false, will append to existing table.
 #' @param tempTable           Should the table created as a temp table?
-#' @param oracleTempSchema    Specifically for Oracle, a schema with write priviliges where temp tables
+#' @param oracleTempSchema    Specifically for Oracle, a schema with write privileges where temp tables
 #'                            can be created.
 #' @param useMppBulkLoad      If using Redshift or PDW, use more performant bulk loading techniques.
 #'                            Setting the system environment variable "USE_MPP_BULK_LOAD" to TRUE is
@@ -272,15 +278,15 @@ insertTable.default <- function(connection,
     tempTable <- TRUE
     warning("Temp table name detected, setting tempTable parameter to TRUE")
   }
-  if (Sys.getenv("USE_MPP_BULK_LOAD") == "TRUE") {
+  if (toupper(Sys.getenv("USE_MPP_BULK_LOAD")) == "TRUE") {
     useMppBulkLoad <- TRUE
   }
   if (dropTableIfExists)
     createTable <- TRUE
   if (tempTable & substr(tableName, 1, 1) != "#" & attr(connection, "dbms") != "redshift")
     tableName <- paste("#", tableName, sep = "")
-  
-  
+
+
   if (is.vector(data) && !is.list(data))
     data <- data.frame(x = data)
   if (length(data) < 1)
@@ -294,7 +300,9 @@ insertTable.default <- function(connection,
     if (!is.data.frame(data))
       data <- as.data.frame(data)
   }
-  
+
+  isSqlReservedWord(c(tableName, colnames(data)), warn = TRUE)
+
   def <- function(obj) {
     if (is.integer(obj)) {
       return("INTEGER")
@@ -323,7 +331,7 @@ insertTable.default <- function(connection,
   fdef <- paste(.sql.qescape(names(data), TRUE, connection@identifierQuote), fts, collapse = ",")
   qname <- .sql.qescape(tableName, TRUE, connection@identifierQuote)
   varNames <- paste(.sql.qescape(names(data), TRUE, connection@identifierQuote), collapse = ",")
-  
+
   if (dropTableIfExists) {
     if (tempTable) {
       sql <- "IF OBJECT_ID('tempdb..@tableName', 'U') IS NOT NULL DROP TABLE @tableName;"
@@ -337,7 +345,7 @@ insertTable.default <- function(connection,
     executeSql(connection, sql, progressBar = FALSE, reportOverallTime = FALSE)
   }
   
-  if (createTable && !tempTable && useMppBulkLoad) {
+  if (createTable && (!tempTable || attr(connection, "dbms") == "hive") && useMppBulkLoad) {
     ensure_installed("uuid")
     ensure_installed("R.utils")
     ensure_installed("urltools")
@@ -345,18 +353,27 @@ insertTable.default <- function(connection,
       stop("MPP credentials could not be confirmed. Please review them or set 'useMppBulkLoad' to FALSE")
     }
     writeLines("Attempting to use MPP bulk loading...")
-    sql <- paste("CREATE TABLE ", qname, " (", fdef, ");", sep = "")
-    sql <- SqlRender::translate(sql, targetDialect = attr(connection, "dbms"))
-    executeSql(connection, sql, progressBar = FALSE, reportOverallTime = FALSE)
+    dbms <- attr(connection, "dbms")
     
-    if (attr(connection, "dbms") == "redshift") {
+    if (dbms != "hive") {
+      sql <- paste("CREATE TABLE ", qname, " (", fdef, ");", sep = "")
+      sql <- SqlRender::translate(sql, targetDialect = attr(connection, "dbms"))
+      executeSql(connection, sql, progressBar = FALSE, reportOverallTime = FALSE)
+    }
+    
+    if (dbms == "redshift") {
       ensure_installed("aws.s3")
       .bulkLoadRedshift(connection, qname, data)
-    } else if (attr(connection, "dbms") == "pdw") {
+    } else if (dbms == "pdw") {
       .bulkLoadPdw(connection, qname, fts, data)
+    } else if (dbms == "hive") {
+      if (tolower(Sys.info()["sysname"]) == "windows") {
+        ensure_installed("ssh")
+      }            
+      .bulkLoadHive(connection, qname, strsplit(varNames, ",")[[1]], data)
     }
   } else {
-    if (attr(connection, "dbms") %in% c("pdw", "redshift", "bigquery") && createTable && nrow(data) > 0) {
+    if (attr(connection, "dbms") %in% c("pdw", "redshift", "bigquery", "hive") && createTable && nrow(data) > 0) {
       ctasHack(connection, qname, tempTable, varNames, fts, data, progressBar, oracleTempSchema)
     } else {
       if (createTable) {
@@ -366,7 +383,7 @@ insertTable.default <- function(connection,
                                     oracleTempSchema = oracleTempSchema)
         executeSql(connection, sql, progressBar = FALSE, reportOverallTime = FALSE)
       }
-      
+
       insertSql <- paste("INSERT INTO ",
                          qname,
                          " (",
@@ -378,13 +395,13 @@ insertTable.default <- function(connection,
       insertSql <- SqlRender::translate(insertSql,
                                         targetDialect = connection@dbms,
                                         oracleTempSchema = oracleTempSchema)
-      
+
       batchSize <- 10000
-      
+
       autoCommit <- rJava::.jcall(connection@jConnection, "Z", "getAutoCommit")
       if (autoCommit) {
-        rJava::.jcall(connection@jConnection, "V", "setAutoCommit", FALSE)
-        on.exit(rJava::.jcall(connection@jConnection, "V", "setAutoCommit", TRUE))
+        trySettingAutoCommit(connection, FALSE)
+        on.exit(trySettingAutoCommit(connection, TRUE))
       }
       if (nrow(data) > 0) {
         if (progressBar) {
@@ -417,7 +434,12 @@ insertTable.default <- function(connection,
             return(NULL)
           }
           lapply(1:ncol(data), setColumn, start = start, end = end)
-          rJava::.jcall(batchedInsert, "V", "executeBatch")
+          if (attr(connection, "dbms") == "bigquery"){
+              rJava::.jcall(batchedInsert, "V", "executeBigQueryBatch")
+          }  else {
+              rJava::.jcall(batchedInsert, "V", "executeBatch")
+          }
+
         }
         if (progressBar) {
           setTxtProgressBar(pb, 1)
@@ -426,6 +448,14 @@ insertTable.default <- function(connection,
       }
     }
   }
+}
+
+trySettingAutoCommit <- function(connection, value) {
+    tryCatch({
+        rJava::.jcall(connection@jConnection, "V", "setAutoCommit", value)
+    }, error = function(cond) {
+        # do nothing
+    })
 }
 
 #' @export
@@ -446,6 +476,8 @@ insertTable.DatabaseConnectorDbiConnection <- function(connection,
     tempTable <- TRUE
     warning("Temp table name detected, setting tempTable parameter to TRUE")
   }
+  isSqlReservedWord(c(tableName, colnames(data)), warn = TRUE)
+
   tableName <- gsub("^#", "", tableName)
   # Convert dates and datetime to UNIX timestamp:
   for (i in 1:ncol(data)) {
@@ -456,7 +488,12 @@ insertTable.DatabaseConnectorDbiConnection <- function(connection,
       data[, i] <- as.numeric(as.POSIXct(data[, i], origin = "1970-01-01", tz = "GMT"))
     }
   }
-  DBI::dbWriteTable(connection@dbiConnection, tableName, data, overwrite = dropTableIfExists, temporary = tempTable)
+  DBI::dbWriteTable(conn = connection@dbiConnection,
+                    name = tableName,
+                    value = data,
+                    overwrite = dropTableIfExists,
+                    append = !createTable,
+                    temporary = tempTable)
   invisible(NULL)
 }
 
@@ -470,23 +507,49 @@ insertTable.DatabaseConnectorDbiConnection <- function(connection,
   } else if (attr(connection, "dbms") == "redshift") {
     envSet <- FALSE
     bucket <- FALSE
-    
+
     if (Sys.getenv("AWS_ACCESS_KEY_ID") != "" && Sys.getenv("AWS_SECRET_ACCESS_KEY") != "" && Sys.getenv("AWS_BUCKET_NAME") !=
         "" && Sys.getenv("AWS_DEFAULT_REGION") != "") {
       envSet <- TRUE
     }
-    
+
     if (aws.s3::bucket_exists(bucket = Sys.getenv("AWS_BUCKET_NAME"))) {
       bucket <- TRUE
     }
-    
+
     if (Sys.getenv("AWS_SSE_TYPE") == "") {
       warning("Not using Server Side Encryption for AWS S3")
     }
     return(envSet & bucket)
+  } else if (attr(connection, "dbms") == "hive") {
+    if (tolower(Sys.info()["sysname"]) == "windows" && !is_installed("ssh")) {
+      writeLines("Please install ssh package, it's required")
+      return(FALSE)
+    }
+    if (Sys.getenv("HIVE_NODE_HOST") == "") {
+      writeLines("Please set environment variable HIVE_NODE_HOST to the Hive Node's host:port")
+      return(FALSE)
+    }
+    if (Sys.getenv("HIVE_SSH_USER") == "") {
+      warning(paste("HIVE_SSH_USER is not set, using default", .getHiveSshUser()))
+    }
+    if (Sys.getenv("HIVE_SSH_PASSWORD") == "" && Sys.getenv("HIVE_KEYFILE") == "") {
+      writeLines("At least one of the following environment variables: HIVE_PASSWORD and/or HIVE_KEYFILE should be set")
+      return(FALSE)
+    }
+    if (Sys.getenv("HIVE_KEYFILE") == "") {
+      warning("Using ssh password authentication, it's recommended to use keyfile instead")
+    }
+    
+    return(TRUE)
   } else {
     return(FALSE)
   }
+}
+
+.getHiveSshUser <- function() {
+  sshUser <- Sys.getenv("HIVE_SSH_USER")
+  return(if (sshUser == "") "root" else sshUser)
 }
 
 .bulkLoadPdw <- function(connection, qname, fts, data) {
@@ -508,22 +571,22 @@ insertTable.DatabaseConnectorDbiConnection <- function(connection,
               sep = "~*~")
   R.utils::gzip(filename = sprintf("%s.csv",
                                    fileName), destname = sprintf("%s.gz", fileName), remove = TRUE)
-  
+
   auth <- sprintf("-U %1s -P %2s", attr(connection, "user"), attr(connection, "password"))
   if (is.null(attr(connection, "user")) && is.null(attr(connection, "password"))) {
     auth <- "-W"
   }
-  
+
   databaseMetaData <- rJava::.jcall(connection@jConnection,
                                     "Ljava/sql/DatabaseMetaData;",
                                     "getMetaData")
   url <- rJava::.jcall(databaseMetaData, "Ljava/lang/String;", "getURL")
   pdwServer <- urltools::url_parse(url)$domain
-  
+
   if (pdwServer == "" | is.null(pdwServer)) {
     stop("PDW Server name cannot be parsed from JDBC URL string")
   }
-  
+
   command <- sprintf("%1s -M append -e UTF8 -i %2s -T %3s -R dwloader.txt -fh 1 -t %4s -r %5s -D ymd -E -se -rv 1 -S %6s %7s",
                      shQuote(Sys.getenv("DWLOADER_PATH")),
                      shQuote(sprintf("%s.gz", fileName)),
@@ -532,7 +595,7 @@ insertTable.DatabaseConnectorDbiConnection <- function(connection,
                      shQuote(eol),
                      pdwServer,
                      auth)
-  
+
   tryCatch({
     system(command,
            intern = FALSE,
@@ -550,7 +613,7 @@ insertTable.DatabaseConnectorDbiConnection <- function(connection,
   }, finally = {
     try(file.remove(sprintf("%s.gz", fileName)), silent = TRUE)
   })
-  
+
   sql <- "SELECT COUNT(*) FROM @table"
   sql <- SqlRender::render(sql, table = qname)
   count <- querySql(connection, sql)
@@ -565,13 +628,13 @@ insertTable.DatabaseConnectorDbiConnection <- function(connection,
   write.csv(x = data, na = "", file = sprintf("%s.csv", fileName), row.names = FALSE, quote = TRUE)
   R.utils::gzip(filename = sprintf("%s.csv",
                                    fileName), destname = sprintf("%s.gz", fileName), remove = TRUE)
-  
+
   s3Put <- aws.s3::put_object(file = sprintf("%s.gz", fileName),
                               check_region = FALSE,
                               headers = list(`x-amz-server-side-encryption` = Sys.getenv("AWS_SSE_TYPE")),
                               object = paste(Sys.getenv("AWS_OBJECT_KEY"), fileName, sep = "/"),
                               bucket = Sys.getenv("AWS_BUCKET_NAME"))
-  
+
   if (!s3Put) {
     stop("Failed to upload data to AWS S3. Please check your credentials and access.")
   }
@@ -584,7 +647,7 @@ insertTable.DatabaseConnectorDbiConnection <- function(connection,
                                            pathToFiles = Sys.getenv("AWS_OBJECT_KEY"),
                                            awsAccessKey = Sys.getenv("AWS_ACCESS_KEY_ID"),
                                            awsSecretAccessKey = Sys.getenv("AWS_SECRET_ACCESS_KEY"))
-  
+
   tryCatch({
     DatabaseConnector::executeSql(connection = connection, sql = sql, reportOverallTime = FALSE)
     delta <- Sys.time() - start
@@ -595,6 +658,65 @@ insertTable.DatabaseConnectorDbiConnection <- function(connection,
     try(file.remove(sprintf("%s.gz", fileName)), silent = TRUE)
     try(aws.s3::delete_object(object = sprintf("%s.gz", fileName),
                               bucket = Sys.getenv("AWS_BUCKET_NAME")), silent = TRUE)
+  })
+}
+
+.bulkLoadHive <- function(connection, qname, varNames, data) {
+  start <- Sys.time()
+  fileName <- sprintf("hive_insert_%s.csv", uuid::UUIDgenerate(use.time = TRUE))
+  filePath <- file.path(tempdir(), fileName)
+  write.csv(x = data, na = "", file = filePath, row.names = FALSE, quote = TRUE)
+  
+  hiveUser <- .getHiveSshUser()
+  hivePasswd <- Sys.getenv("HIVE_SSH_PASSWORD")
+  hiveHost <- Sys.getenv("HIVE_NODE_HOST")
+  sshPort <- (function(port) if (port == "") "2222" else port)(Sys.getenv("HIVE_SSH_PORT"))
+  nodePort <- (function(port) if (port == "") "8020" else port)(Sys.getenv("HIVE_NODE_PORT"))
+  hiveKeyFile <- (function(keyfile) if (keyfile == "") NULL else keyfile)(Sys.getenv("HIVE_KEYFILE"))
+  hadoopUser <- (function(hadoopUser) if (hadoopUser == "") "hive" else hadoopUser)(Sys.getenv("HADOOP_USER_NAME"))
+  
+  tryCatch({
+    if (tolower(Sys.info()["sysname"]) == "windows") {
+      session <- ssh::ssh_connect(host = sprintf("%s@%s:%s", hiveUser, hiveHost, sshPort), passwd = hivePasswd, keyfile = hiveKeyFile)
+      remoteFile <- paste0("/tmp/", fileName)
+      ssh::scp_upload(session, filePath, to = remoteFile, verbose = FALSE)
+      hadoopDir <- sprintf("/user/%s/%s", hadoopUser, uuid::UUIDgenerate(use.time = TRUE))
+      hadoopFile <- paste0(hadoopDir, "/", fileName)
+      ssh::ssh_exec_wait(session, sprintf("HADOOP_USER_NAME=%s hadoop fs -mkdir %s", hadoopUser, hadoopDir))
+      command <- sprintf("HADOOP_USER_NAME=%s hadoop fs -put %s %s", hadoopUser, remoteFile, hadoopFile)
+      ssh::ssh_exec_wait(session, command = command)
+    } else {
+      remoteFile <- paste0("/tmp/", fileName)
+      scp_command <- sprintf("sshpass -p \'%s\' scp -P %s %s %s:%s", hivePasswd, sshPort, filePath, hiveHost, remoteFile)
+      system(scp_command)
+      hadoopDir <- sprintf("/user/%s/%s", hadoopUser, uuid::UUIDgenerate(use.time = TRUE))
+      hadoopFile <- paste0(hadoopDir, "/", fileName)
+      hdp_mk_dir_command <- sprintf("sshpass -p \'%s\' ssh %s -p %s HADOOP_USER_NAME=%s hadoop fs -mkdir %s", hivePasswd, hiveHost, sshPort, hadoopUser, hadoopDir)
+      system(hdp_mk_dir_command)
+      hdp_put_command <- sprintf("sshpass -p \'%s\' ssh %s -p %s HADOOP_USER_NAME=%s hadoop fs -put %s %s", hivePasswd, hiveHost, sshPort, hadoopUser, remoteFile, hadoopFile)
+      system(hdp_put_command)
+    }
+    def <- function(name) {
+      return(paste(name, "STRING"))
+    }
+    fdef <- paste(sapply(varNames, def), collapse = ", ")
+    sql <- SqlRender::render("CREATE TABLE @table(@fdef) ROW FORMAT DELIMITED FIELDS TERMINATED BY ',' STORED AS TEXTFILE LOCATION 'hdfs://@hiveHost:@nodePort@filename';", 
+                             filename = hadoopDir, table = qname, fdef = fdef, hiveHost = hiveHost, nodePort = nodePort)
+    sql <- SqlRender::translate(sql, targetDialect = "hive", oracleTempSchema = NULL)
+    
+    tryCatch({
+      DatabaseConnector::executeSql(connection = connection, sql = sql, reportOverallTime = FALSE)
+      delta <- Sys.time() - start
+      writeLines(paste("Bulk load to Hive took", signif(delta, 3), attr(delta, "units")))
+    }, error = function(e) {
+      writeLines(paste("Error:", e))
+      stop("Error in Hive bulk upload.")
+    }, finally = {
+      try(invisible(file.remove(filePath)))
+    })
+  }, finally = {
+    if (tolower(Sys.info()["sysname"]) == "windows")
+      ssh::ssh_disconnect(session)
   })
 }