+++ inst/doc/UsingDatabaseConnector.R
@@ -1,4 +1,4 @@
-## ---- echo = FALSE, message = FALSE--------------------------------------
+## ---- echo = FALSE, message = FALSE-------------------------------------------
 library(DatabaseConnector)
 knitr::opts_chunk$set(
   cache=FALSE,
@@ -6,33 +6,33 @@ knitr::opts_chunk$set(
   error = FALSE,
   tidy = FALSE)
 
-## ----eval=FALSE----------------------------------------------------------
+## ----eval=FALSE---------------------------------------------------------------
 #  ?jdbcDrivers
 
-## ----eval=FALSE----------------------------------------------------------
+## ----eval=FALSE---------------------------------------------------------------
 #  conn <- connect(dbms = "postgresql",
 #                  server = "localhost/postgres",
 #                  user = "joe",
 #                  password = "secret",
 #                  schema = "cdm")
 
-## ----echo=FALSE----------------------------------------------------------
+## ----echo=FALSE---------------------------------------------------------------
 writeLines("Connecting using PostgreSQL driver")
 
-## ----eval=FALSE----------------------------------------------------------
+## ----eval=FALSE---------------------------------------------------------------
 #  disconnect(conn)
 
-## ----eval=FALSE----------------------------------------------------------
+## ----eval=FALSE---------------------------------------------------------------
 #  conn <- connect(dbms = "postgresql",
 #                  connectionString = "jdbc:postgresql://localhost:5432/postgres",
 #                  user = "joe",
 #                  password = "secret",
 #                  schema = "cdm")
 
-## ----echo=FALSE----------------------------------------------------------
+## ----echo=FALSE---------------------------------------------------------------
 writeLines("Connecting using PostgreSQL driver")
 
-## ----eval=FALSE----------------------------------------------------------
+## ----eval=FALSE---------------------------------------------------------------
 #  details <- createConnectionDetails(dbms = "postgresql",
 #                                     server = "localhost/postgres",
 #                                     user = "joe",
@@ -40,10 +40,10 @@ writeLines("Connecting using PostgreSQL driver")
 #                                     schema = "cdm")
 #  conn <- connect(details)
 
-## ----echo=FALSE----------------------------------------------------------
+## ----echo=FALSE---------------------------------------------------------------
 writeLines("Connecting using PostgreSQL driver")
 
-## ----eval=FALSE----------------------------------------------------------
+## ----eval=FALSE---------------------------------------------------------------
 #  details <- createConnectionDetails(dbms = "netezza",
 #                                     server = "myserver.com/mainDb",
 #                                     user = "joe",
@@ -52,31 +52,33 @@ writeLines("Connecting using PostgreSQL driver")
 #                                     pathToDriver = "c:/temp")
 #  conn <- connect(details)
 
-## ----echo=FALSE----------------------------------------------------------
+## ----echo=FALSE---------------------------------------------------------------
 writeLines("Connecting using Netezza driver")
 
-## ----eval=FALSE----------------------------------------------------------
+## ----eval=FALSE---------------------------------------------------------------
 #  querySql(conn, "SELECT TOP 3 * FROM person")
 
-## ----echo=FALSE----------------------------------------------------------
+## ----echo=FALSE---------------------------------------------------------------
 data.frame(PERSON_ID = c(1,2,3), GENDER_CONCEPT_ID = c(8507, 8507, 8507), YEAR_OF_BIRTH = c(1975, 1976, 1977))
 
-## ----eval=FALSE----------------------------------------------------------
+## ----eval=FALSE---------------------------------------------------------------
 #  executeSql(conn, "TRUNCATE TABLE foo; DROP TABLE foo; CREATE TABLE foo (bar INT);")
 
-## ----eval=FALSE----------------------------------------------------------
-#  x <- querySql.ffdf(conn, "SELECT * FROM person")
+## ----eval=FALSE---------------------------------------------------------------
+#  library(Andromeda)
+#  x <- andromeda()
+#  querySqlToAndromeda(conn, "SELECT * FROM person", andromeda = x, andromedaTableName = "person)
 
-## ----eval=FALSE----------------------------------------------------------
+## ----eval=FALSE---------------------------------------------------------------
 #  persons <- renderTranslatequerySql(conn,
 #                                     sql = "SELECT TOP 10 * FROM @schema.person",
 #                                     schema = "cdm_synpuf")
 
-## ----eval=FALSE----------------------------------------------------------
+## ----eval=FALSE---------------------------------------------------------------
 #  data(mtcars)
 #  insertTable(conn, "mtcars", mtcars, createTable = TRUE)
 
-## ----eval=FALSE----------------------------------------------------------
+## ----eval=FALSE---------------------------------------------------------------
 #  conn <- dbConnect(DatabaseConnectorDriver(),
 #                    dbms = "postgresql",
 #                    server = "localhost/postgres",
@@ -84,33 +86,33 @@ data.frame(PERSON_ID = c(1,2,3), GENDER_CONCEPT_ID = c(8507, 8507, 8507), YEAR_O
 #                    password = "secret",
 #                    schema = "cdm")
 
-## ----echo=FALSE----------------------------------------------------------
+## ----echo=FALSE---------------------------------------------------------------
 writeLines("Connecting using PostgreSQL driver")
 
-## ----eval=FALSE----------------------------------------------------------
+## ----eval=FALSE---------------------------------------------------------------
 #  dbIsValid(conn)
 
-## ----echo=FALSE----------------------------------------------------------
+## ----echo=FALSE---------------------------------------------------------------
 TRUE
 
-## ----eval=FALSE----------------------------------------------------------
+## ----eval=FALSE---------------------------------------------------------------
 #  res <- dbSendQuery(conn, "SELECT TOP 3 * FROM person")
 #  dbFetch(res)
 
-## ----echo=FALSE----------------------------------------------------------
+## ----echo=FALSE---------------------------------------------------------------
 data.frame(PERSON_ID = c(1,2,3), GENDER_CONCEPT_ID = c(8507, 8507, 8507), YEAR_OF_BIRTH = c(1975, 1976, 1977))
 
-## ----eval=FALSE----------------------------------------------------------
+## ----eval=FALSE---------------------------------------------------------------
 #  dbHasCompleted(res)
 
-## ----echo=FALSE----------------------------------------------------------
+## ----echo=FALSE---------------------------------------------------------------
 TRUE
 
-## ----eval=FALSE----------------------------------------------------------
+## ----eval=FALSE---------------------------------------------------------------
 #  dbClearResult(res)
 #  dbDisconnect(res)
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 conn <- connect(dbms = "sqlite", server = tempfile())
 
 # Upload cars dataset as table: